pyproject.toml
[project]
name = "youtube-shorts-bot"
version = "0.1.0"
description = "Telegram bot for downloading YouTube Shorts with queue management"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot>=21.0",
    "yt-dlp>=2024.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
